MPO Representation

As a quantum state can be converted into MPS states, an arbitrary quantum operation can similarly be converted into a Matrix-Product Operator (MPO) form. MPO objects can be created from an AbtractMatrix or a given CircuitGate object.

Qaintensor.MPO — Type

MPO(tensors::AbstractVector{Tensor}, contractions::AbstractVector{Summation}, openidx::AbstractVector{Pair{T,T}}) where T <: Integer

Subclassed TensorNetwork object in Matrix Product Operator(MPO) form.

Qaintensor.MPO — Method

MPO(m::AbstractMatrix)

Transform an operator represented by matrix m into an MPO form.

Qaintensor.MPO — Method

MPO(cg::CircuitGate)

Transform an operator represented by a gate cg into an MPO form.

Qaintensor.MPO — Method

MPO(cg::CircuitGate)

Transform an operator represented by a gate cg into an MPO form.

MPO objects can be applied to TensorNetwork objects via the apply_MPO function prior to contraction, given a tuple of wires the MPO is applied to.

Qaintensor.apply_MPO — Method

apply_MPO(ψ::TensorNetwork, mpo::MPO, iwire::NTuple{M, <:Integer}) where {M}

Given a state ψ in a Tensor Network form and an operator mpo acting on M qudits, update the state by effectively applying mpo.

Arguments

iwire::NTuple{M, <:Integer}: qudits in which MPO acts. When the input is mpo::MPO, iwires must be sorted.