MPS Representation

One of the most common specialized tensor networks is the Matrix-Product State (MPS). This can be constructed from a given quantum state in state vector form via a series of SVD contractions.

Qaintensor.jl currently supports the generation of a general MPS objet from a given quantum state.

Qaintensor.MPS — Method

MPS(ψ::AbstractVector{ComplexF64})

Convert a vector ψ into a MPS reprsentation of it.

Specialized MPS forms, such as OpenMPS or ClosedMPS are also supported.

Qaintensor.OpenMPS — Method

OpenMPS(T::AbstractVector{Tensor})

Create a MPS formed by the tensors in T with open boundary conditions (that is, each tensor has 2 virtual legs and a physical one, including those on the boundaries).

Qaintensor.ClosedMPS — Method

ClosedMPS(T::AbstractVector{Tensor})

Create a MPS formed by the tensors in T with closed boundary conditions (that is, each tensor has 2 virtual legs and a physical one, except those at the boundaries that have only one virtual leg).

MPS objects can be contracted via the contract function (see TensorNetwork Contractions).