Tensor Networks

The a tensor network consists of multiple tensors connected by bonds. The contraction of these bonds determines the final output of the tensor network. Quantum Circuits are a specialized subset of tensor networks.

Tensors

The basic building block of any GeneralTensorNetwork is a Tensor. These are constructed from an arbitrary tensor or high-order AbstractMatrix.

Qaintensor.Tensor — Type

Tensor

stores vector of tensors

GeneralTensorNetwork

All tensor network objects are based off the abstract TensorNetwork struct. The most general form is the GeneralTensorNetwork. This object can be created by providing a list of Tensors objects, contractions sequences and open indices.

Qaintensor.GeneralTensorNetwork — Type

GeneralTensorNetwork  <: TensorNetwork

General Tensor network, consisting of tensors and contraction operations

TensorNetwork objects can be contracted via the contract function.

TensorNetwork Contractions

TensorNetworks can be contracted via the contract function.

Qaintensor.contract — Method

contract(net::TensorNetwork; optimize=false)

Fully contract a given TensorNetwork object.